move useful functions to utils
